requirements.txt
einops
omegaconf
safetensors
albumentations==0.4.3
diffusers
transformers
opencv-python==4.3.0.38
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
test-tube>=0.7.5
streamlit>=0.73.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
kornia==0.6
streamlit-drawable-canvas==0.8
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e git+https://github.com/openai/CLIP.git@main#egg=clip